Rename image.py to image_mockup.py
